TapBar ContentView에 넣어뒀습니다.
